Split sycl_iterator_for.pass.cpp to set of separate tests
