+++ test/general/sycl_iterator/sycl_iterator_for_test_copy.pass.cpp
@@ -1,0 +1,76 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Copyright (C) Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+// This file incorporates work covered by the following copyright and permission
+// notice:
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+//
+//===----------------------------------------------------------------------===//
+
+#include "sycl_iterator_test.h"
+
+#if TEST_DPCPP_BACKEND_PRESENT
+
+DEFINE_TEST(test_copy)
+{
+    DEFINE_TEST_CONSTRUCTOR(test_copy, 1.0f, 1.0f)
+
+    template <typename Policy, typename Iterator1, typename Iterator2, typename Size>
+    void
+    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 /* last2 */, Size n)
+    {
+        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
+        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, n);
+
+        using IteratorValueType = typename ::std::iterator_traits<Iterator1>::value_type;
+        auto value = IteratorValueType(42);
+        ::std::fill(host_keys.get(), host_keys.get() + n, value);
+        ::std::fill(host_vals.get(), host_vals.get() + n, IteratorValueType{0});
+        update_data(host_keys, host_vals);
+
+        ::std::copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
+        wait_and_throw(exec);
+
+        host_vals.retrieve_data();
+        EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
+                    "wrong effect from copy");
+    }
+};
+
+template <sycl::usm::alloc alloc_type>
+void
+test_usm_and_buffer()
+{
+    using ValueType = ::std::int32_t;
+
+    PRINT_DEBUG("test_copy");
+    test2buffers<alloc_type, test_copy<ValueType>>();
+}
+#endif // TEST_DPCPP_BACKEND_PRESENT
+
+std::int32_t
+main()
+{
+    try
+    {
+#if TEST_DPCPP_BACKEND_PRESENT
+        // Run tests for USM shared memory
+        test_usm_and_buffer<sycl::usm::alloc::shared>();
+        // Run tests for USM device memory
+        test_usm_and_buffer<sycl::usm::alloc::device>();
+#endif // TEST_DPCPP_BACKEND_PRESENT
+    }
+    catch (const ::std::exception& exc)
+    {
+        std::cout << "Exception: " << exc.what() << std::endl;
+        return EXIT_FAILURE;
+    }
+
+    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+}
