+++ test/general/sycl_iterator/sycl_iterator_for_test_destroy_n.pass.cpp
@@ -1,0 +1,103 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Copyright (C) Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+// This file incorporates work covered by the following copyright and permission
+// notice:
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+//
+//===----------------------------------------------------------------------===//
+
+#include "sycl_iterator_test.h"
+
+#if TEST_DPCPP_BACKEND_PRESENT
+
+// created just to check destroy and destroy_n correctness
+template <typename T>
+struct SyclTypeWrapper
+{
+    T __value;
+
+    explicit SyclTypeWrapper(const T& value = T{4}) : __value(value) {}
+    ~SyclTypeWrapper() { __value = -2; }
+    bool
+    operator==(const SyclTypeWrapper& other) const
+    {
+        return __value == other.__value;
+    }
+};
+
+// this wrapper is needed to take into account not only kernel name,
+// but also other types (for example, iterator's value type)
+template <typename... T>
+struct policy_name_wrapper
+{
+};
+
+DEFINE_TEST(test_destroy_n)
+{
+    DEFINE_TEST_CONSTRUCTOR(test_destroy_n, 1.0f, 1.0f)
+
+    template <typename Policy, typename Iterator1, typename Size>
+    void
+    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
+    {
+        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
+
+        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
+        auto value = T1{ 2 };
+
+        ::std::fill(host_keys.get(), host_keys.get() + n, value);
+        host_keys.update_data();
+
+        ::std::destroy_n(make_new_policy<policy_name_wrapper<new_kernel_name<Policy, 0>, T1>>(exec), first1, n);
+        if(!::std::is_trivially_destructible_v<T1>)
+            value = T1{-2};
+        wait_and_throw(exec);
+
+        host_keys.retrieve_data();
+        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value),
+                    "wrong effect from destroy_n");
+    }
+};
+
+using ValueType = ::std::int32_t;
+
+template<>
+struct sycl::is_device_copyable<SyclTypeWrapper<ValueType>> : std::true_type {};
+
+template <sycl::usm::alloc alloc_type>
+void
+test_usm_and_buffer()
+{
+    PRINT_DEBUG("test_destroy_n");
+    test1buffer<alloc_type, test_destroy_n<SyclTypeWrapper<ValueType>>>();
+    test1buffer<alloc_type, test_destroy_n<ValueType>>();
+}
+#endif // TEST_DPCPP_BACKEND_PRESENT
+
+std::int32_t
+main()
+{
+    try
+    {
+#if TEST_DPCPP_BACKEND_PRESENT
+        // Run tests for USM shared memory
+        test_usm_and_buffer<sycl::usm::alloc::shared>();
+        // Run tests for USM device memory
+        test_usm_and_buffer<sycl::usm::alloc::device>();
+#endif // TEST_DPCPP_BACKEND_PRESENT
+    }
+    catch (const ::std::exception& exc)
+    {
+        std::cout << "Exception: " << exc.what() << std::endl;
+        return EXIT_FAILURE;
+    }
+
+    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+}
