+++ test/general/sycl_iterator/sycl_iterator_for_test_for_each.pass.cpp
@@ -1,0 +1,90 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Copyright (C) Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+// This file incorporates work covered by the following copyright and permission
+// notice:
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+//
+//===----------------------------------------------------------------------===//
+
+#include "sycl_iterator_test.h"
+
+#if TEST_DPCPP_BACKEND_PRESENT
+
+struct Inc
+{
+    template <typename T>
+    void
+    operator()(T& x) const
+    {
+        ++x;
+    }
+};
+
+DEFINE_TEST(test_for_each)
+{
+    DEFINE_TEST_CONSTRUCTOR(test_for_each, 1.0f, 1.0f)
+
+    template <typename Policy, typename Iterator1, typename Size>
+    void
+    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
+    {
+        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
+
+        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
+        auto value = T1(6);
+
+        ::std::fill(host_keys.get(), host_keys.get() + n, value);
+        ::std::fill(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value - 1);
+        host_keys.update_data();
+
+        ::std::for_each(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2), Inc());
+        wait_and_throw(exec);
+
+        // We call due to SYCL 1.2.1: 4.7.2.3.
+        // If the host memory is modified by the host,
+        // or mapped to another buffer or image during the lifetime of this buffer,
+        // then the results are undefined
+        host_keys.retrieve_data();
+        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value), "wrong effect from for_each");
+    }
+};
+
+template <sycl::usm::alloc alloc_type>
+void
+test_usm_and_buffer()
+{
+    using ValueType = ::std::int32_t;
+
+    // test1buffer
+    PRINT_DEBUG("test_for_each");
+    test1buffer<alloc_type, test_for_each<ValueType>>();
+}
+#endif // TEST_DPCPP_BACKEND_PRESENT
+
+std::int32_t
+main()
+{
+    try
+    {
+#if TEST_DPCPP_BACKEND_PRESENT
+        // Run tests for USM shared memory
+        test_usm_and_buffer<sycl::usm::alloc::shared>();
+        // Run tests for USM device memory
+        test_usm_and_buffer<sycl::usm::alloc::device>();
+#endif // TEST_DPCPP_BACKEND_PRESENT
+    }
+    catch (const ::std::exception& exc)
+    {
+        std::cout << "Exception: " << exc.what() << std::endl;
+        return EXIT_FAILURE;
+    }
+
+    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+}
