+++ test/general/sycl_iterator/sycl_iterator_for_test_generate.pass.cpp
@@ -1,0 +1,90 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Copyright (C) Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+// This file incorporates work covered by the following copyright and permission
+// notice:
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+//
+//===----------------------------------------------------------------------===//
+
+#include "sycl_iterator_test.h"
+
+#if TEST_DPCPP_BACKEND_PRESENT
+
+template <typename T>
+struct Generator_count
+{
+    T def_val;
+    Generator_count(const T& val) : def_val(val) {}
+    T
+    operator()() const
+    {
+        return def_val;
+    }
+    T
+    default_value() const
+    {
+        return def_val;
+    }
+};
+
+DEFINE_TEST(test_generate)
+{
+    DEFINE_TEST_CONSTRUCTOR(test_generate, 1.0f, 1.0f)
+
+    template <typename Policy, typename Iterator1, typename Size>
+    void
+    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
+    {
+        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
+
+        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
+        auto value = T1(4);
+
+        ::std::generate(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2),
+                      Generator_count<T1>(value));
+        wait_and_throw(exec);
+
+        host_keys.retrieve_data();
+        EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value),
+                    "wrong effect from generate");
+    }
+};
+
+template <sycl::usm::alloc alloc_type>
+void
+test_usm_and_buffer()
+{
+    using ValueType = ::std::int32_t;
+
+    PRINT_DEBUG("test_generate");
+    test1buffer<alloc_type, test_generate<ValueType>>();
+}
+#endif // TEST_DPCPP_BACKEND_PRESENT
+
+std::int32_t
+main()
+{
+    try
+    {
+#if TEST_DPCPP_BACKEND_PRESENT
+        // Run tests for USM shared memory
+        test_usm_and_buffer<sycl::usm::alloc::shared>();
+        // Run tests for USM device memory
+        test_usm_and_buffer<sycl::usm::alloc::device>();
+#endif // TEST_DPCPP_BACKEND_PRESENT
+    }
+    catch (const ::std::exception& exc)
+    {
+        std::cout << "Exception: " << exc.what() << std::endl;
+        return EXIT_FAILURE;
+    }
+
+    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+}
