+++ test/general/sycl_iterator/sycl_iterator_for_test_uninitialized_default_construct.pass.cpp
@@ -1,0 +1,97 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Copyright (C) Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+// This file incorporates work covered by the following copyright and permission
+// notice:
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+//
+//===----------------------------------------------------------------------===//
+
+#include "sycl_iterator_test.h"
+
+#if TEST_DPCPP_BACKEND_PRESENT
+
+// created just to check destroy and destroy_n correctness
+template <typename T>
+struct SyclTypeWrapper
+{
+    T __value;
+
+    explicit SyclTypeWrapper(const T& value = T{4}) : __value(value) {}
+    ~SyclTypeWrapper() { __value = -2; }
+    bool
+    operator==(const SyclTypeWrapper& other) const
+    {
+        return __value == other.__value;
+    }
+};
+
+DEFINE_TEST(test_uninitialized_default_construct)
+{
+    DEFINE_TEST_CONSTRUCTOR(test_uninitialized_default_construct, 1.0f, 1.0f)
+
+    template <typename Policy, typename Iterator1, typename Size>
+    void
+    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
+    {
+        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
+
+        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
+        auto value = T1{ 2 };
+
+        T1 exp_value; // default-constructed value
+        ::std::fill(host_keys.get(), host_keys.get() + n, value);
+        host_keys.update_data();
+
+        ::std::uninitialized_default_construct(make_new_policy<new_kernel_name<Policy, 0>>(exec),
+                                             first1 + (n / 3), first1 + (n / 2));
+        wait_and_throw(exec);
+
+        host_keys.retrieve_data();
+        EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), exp_value),
+                    "wrong effect from uninitialized_default_construct");
+    }
+};
+
+using ValueType = ::std::int32_t;
+
+template <>
+struct sycl::is_device_copyable<SyclTypeWrapper<ValueType>> : std::true_type
+{
+};
+
+template <sycl::usm::alloc alloc_type>
+void
+test_usm_and_buffer()
+{
+    PRINT_DEBUG("test_uninitialized_default_construct");
+    test1buffer<alloc_type, test_uninitialized_default_construct<SyclTypeWrapper<ValueType>>>();
+}
+#endif // TEST_DPCPP_BACKEND_PRESENT
+
+std::int32_t
+main()
+{
+    try
+    {
+#if TEST_DPCPP_BACKEND_PRESENT
+        // Run tests for USM shared memory
+        test_usm_and_buffer<sycl::usm::alloc::shared>();
+        // Run tests for USM device memory
+        test_usm_and_buffer<sycl::usm::alloc::device>();
+#endif // TEST_DPCPP_BACKEND_PRESENT
+    }
+    catch (const ::std::exception& exc)
+    {
+        std::cout << "Exception: " << exc.what() << std::endl;
+        return EXIT_FAILURE;
+    }
+
+    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
+}
